+++ CHANGELOG.md
@@ -189,6 +189,8 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
 - ADGroup
   - Refactored Module.
   - Refactored Unit and Integration Tests.
+- ADUser
+  - Added SamAccountName property.
 
 ### Added